% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{add_plural_rule}
\alias{add_plural_rule}
\title{Add a custom rule for making a word plural}
\usage{
add_plural_rule(singular_word, plural_word)
}
\arguments{
\item{singular_word}{desired singular form of a word}

\item{plural_word}{plural form of a word}
}
\description{
Add a custom rule for making a word plural
}
\examples{
add_plural_rule("gex", "gexii")
pluralize('regex')
singularize('singles')
}
