% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{add_singular_rule}
\alias{add_singular_rule}
\title{Add a custom rule for making a word singular}
\usage{
add_singular_rule(plural_word, singular_word)
}
\arguments{
\item{plural_word}{plural form of a word}

\item{singular_word}{desired singular form of a word}
}
\value{
nothing (function produces a side effect of modifying
        in-memory internal package envrionment)
}
\description{
Add a custom rule for making a word singular
}
\examples{
add_singular_rule("singles", "singular")
singularize('singles')
pluralize("irregular")
pluralize(c("woman", "man", "child", "tooth", "foot", "person", "leaf"))
singularize(c("woman", "man", "child", "tooth", "foot", "person", "leaf"))
}
