% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize.r
\name{add_uncountable_rule}
\alias{add_uncountable_rule}
\title{Make a word "uncountable"}
\usage{
add_uncountable_rule(uncountable_word)
}
\description{
Sometimes you don't want to turn "\code{paper}" into "\code{papers}". You
can use this function to make a word "uncountable".
}
\examples{
pluralize('test')
singularize('test')
pluralize('test', 5)
pluralize('test', 5, TRUE)
pluralize('regex')
add_plural_rule("gex", "gexii")
pluralize('regex')
singularize('singles')
add_singular_rule("singles", "singular")
singularize('singles')
pluralize("irregular")
add_irregular_rule("irregular", "regular")
pluralize("irregular")
pluralize("paper")
add_uncountable_rule("paper")
pluralize("paper")
}

