% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_plural}
\alias{is_plural}
\alias{is_singular}
\title{Test plural state of a word}
\usage{
is_plural(x)

is_singular(x)
}
\arguments{
\item{x}{vector of words to test}
}
\value{
logical vector
}
\description{
Test plural state of a word
}
\examples{
is_singular(c("boats", "house", "cats", "river"))
is_plural(c("boats", "house", "cats", "river"))
}
