% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize-package.R
\docType{package}
\name{pluralize-package}
\alias{pluralize-package}
\title{Pluralize and Singularize Any Word}
\description{
Tools are provided to create plural, singular and regular forms of
English words along with tools to augment the built-in rules to fit
specializied needs. Core functionality is based on a JavaScript library,
<https://github.com/blakeembrey/pluralize>, by Blake Embrey.
}
\author{
Bob Rudis (@hrbrmstr)
}
\keyword{internal}
