% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize.r
\name{singularize}
\alias{singularize}
\title{Singularize a word}
\usage{
singularize(x)
}
\arguments{
\item{x}{vector of words to make singular}
}
\description{
If the built-in rules are not sufficient, use \code{add_singular_rule()} to
customize the behavior.
}
\examples{
pluralize('test')
singularize('test')
pluralize('test', 5)
pluralize('test', 5, TRUE)
pluralize('regex')
add_plural_rule("gex", "gexii")
pluralize('regex')
singularize('singles')
add_singular_rule("singles", "singular")
singularize('singles')
pluralize("irregular")
add_irregular_rule("irregular", "regular")
pluralize("irregular")
pluralize("paper")
add_uncountable_rule("paper")
pluralize("paper")
}

