% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize.r
\name{add_irregular_rule}
\alias{add_irregular_rule}
\title{Add a custom rule for making "deregularizing" a word}
\usage{
add_irregular_rule(irregular_word, regular_word)
}
\arguments{
\item{singular_word}{desired singular form of a word}

\item{plural_word}{plural form of a word}
}
\description{
"\code{regular}" is the "deregularized" form of "\code{irregular}".
Calling \code{plural("irregular")} will return "\code{regular}" if you
define a rule such as \code{add_irregular_rule("irregular", "regular")}.
}
\examples{
pluralize('test')
singularize('test')
pluralize('test', 5)
pluralize('test', 5, TRUE)
pluralize('regex')
add_plural_rule("gex", "gexii")
pluralize('regex')
singularize('singles', 1)
add_singular_rule("singles", "singular")
singularize('singles', 1)
pluralize("irregular")
add_irregular_rule("irregular", "regular")
pluralize("irregular")
pluralize("paper")
add_uncountable_rule("paper")
pluralize("paper")
}

