% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{add_irregular_rule}
\alias{add_irregular_rule}
\title{Add a custom rule for making "deregularizing" a word}
\usage{
add_irregular_rule(irregular_word, regular_word)
}
\arguments{
\item{irregular_word}{the irregular form of the word}

\item{regular_word}{regular form of a word}
}
\value{
nothing (function produces a side effect of modifying
        in-memory internal package envrionment)
}
\description{
"\code{regular}" is the "deregularized" form of "\code{irregular}".
Calling \code{plural("irregular")} will return "\code{regular}" if you
define a rule such as \code{add_irregular_rule("irregular", "regular")}.
}
\examples{
add_irregular_rule("irregular", "regular")
pluralize("irregular")
pluralize("paper")
}
