% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize.r
\name{add_plural_rule}
\alias{add_plural_rule}
\title{Add a custom rule for making a word plural}
\usage{
add_plural_rule(singular_word, plural_word)
}
\arguments{
\item{singular_word}{desired singular form of a word}

\item{plural_word}{plural form of a word}
}
\description{
Add a custom rule for making a word plural
}
\examples{
pluralize('test')
singularize('test')
pluralize('test', 5)
pluralize('test', 5, TRUE)
pluralize('regex')
add_plural_rule("gex", "gexii")
pluralize('regex')
singularize('singles')
add_singular_rule("singles", "singular")
singularize('singles')
pluralize("irregular")
add_irregular_rule("irregular", "regular")
pluralize("irregular")
pluralize("paper")
add_uncountable_rule("paper")
pluralize("paper")
}

