% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize.r
\name{add_uncountable_rule}
\alias{add_uncountable_rule}
\title{Make a word "uncountable"}
\usage{
add_uncountable_rule(uncountable_word)
}
\arguments{
\item{uncountable_word}{the word that's, well, uncountable}
}
\description{
Sometimes you don't want to turn "\code{paper}" into "\code{papers}". You
can use this function to make a word "uncountable".
}
\examples{
pluralize('test')
singularize('test')
singularize(c("boats", "houses", "cats", "rivers"))
pluralize(singularize(c("boats", "houses", "cats", "rivers")))
singularize(c("buses", "wishes", "pitches", "boxexs"))
pluralize(singularize(c("buses", "wishes", "pitches", "boxexs")))
singularize(c("pennies", "spies", "babies", "cities", "daisies"))
pluralize(singularize(c("pennies", "spies", "babies", "cities", "daisies")))
singularize(c("sheep", "fish", "deer", "species", "aircraft"))
pluralize(singularize(c("sheep", "fish", "deer", "species", "aircraft")))
pluralize('test', 5)
pluralize('test', 5, TRUE)
pluralize('regex')
add_plural_rule("gex", "gexii")
pluralize('regex')
singularize('singles')
add_singular_rule("singles", "singular")
singularize('singles')
pluralize("irregular")
pluralize(c("woman", "man", "child", "tooth", "foot", "person", "leaf"))
singularize(c("woman", "man", "child", "tooth", "foot", "person", "leaf"))
add_irregular_rule("irregular", "regular")
pluralize("irregular")
pluralize("paper")
add_uncountable_rule("paper")
pluralize("paper")
}

