% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{add_uncountable_rule}
\alias{add_uncountable_rule}
\title{Make a word "uncountable"}
\usage{
add_uncountable_rule(uncountable_word)
}
\arguments{
\item{uncountable_word}{the word that's, well, uncountable}
}
\value{
nothing (function produces a side effect of modifying
        in-memory internal package envrionment)
}
\description{
Sometimes you don't want to turn "\code{paper}" into "\code{papers}". You
can use this function to make a word "uncountable".
}
\examples{
add_uncountable_rule("paper")
pluralize("paper")
}
