% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize-package.R, R/pluralize.r
\docType{package}
\name{pluralize}
\alias{pluralize}
\alias{pluralize-package}
\title{Pluralize and Singularize Any Word}
\usage{
pluralize(x, n = 2, prepend = FALSE)
}
\arguments{
\item{x}{character vector of words to make plural}

\item{n}{amount of plural (some plurals change by amount and this also impacts
the \code{prepend} functionality)}

\item{prepend}{should we prepend the value of \code{n} to the output? Default: \code{FALSE}}
}
\value{
character vector of pluralized words
}
\description{
Based on the \href{https://github.com/blakeembrey/pluralize}{pluralize.js}
javascript library by \code{Blake Embrey}.

If the built-in rules are not sufficient, use \code{add_plural_rule()} to
customize the behavior.
}
\examples{
pluralize('test')
singularize('test')
pluralize('test', 5)
pluralize('test', 5, TRUE)
pluralize('regex')
add_plural_rule("gex", "gexii")
pluralize('regex')
singularize('singles')
add_singular_rule("singles", "singular")
singularize('singles')
pluralize("irregular")
add_irregular_rule("irregular", "regular")
pluralize("irregular")
pluralize("paper")
add_uncountable_rule("paper")
pluralize()
}
\author{
Bob Rudis (@hrbrmstr)
}

