% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize-package.R, R/pluralize.R
\docType{package}
\name{pluralize}
\alias{pluralize}
\title{Pluralize and Singularize Any Word}
\usage{
pluralize(x, n = 2, prepend = FALSE)
}
\arguments{
\item{x}{character vector of words to make plural}

\item{n}{amount of plural (some plurals change by amount and this also impacts
the \code{prepend} functionality)}

\item{prepend}{should we prepend the value of \code{n} to the output? Default: \code{FALSE}}
}
\value{
character vector of modified pluralized words
}
\description{
Tools are provided to create plural, singular and regular forms of
English words along with tools to augment the built-in rules to fit
specialzied needs. Core functionality is based on a JavaScript library,
<https://github.com/blakeembrey/pluralize>, by Blake Embrey.

If the built-in rules are not sufficient, use \code{add_plural_rule()} to
customize the behavior.
}
\examples{
pluralize('test')
pluralize('test', 5)
pluralize('test', 5, TRUE)
pluralize('regex')
}
\author{
Bob Rudis (@hrbrmstr)
}
